fix: chain description yield now matches the table stats row
The description box computed its 30-day average yield over all chain
vaults (2.4% on Base) while the stats row below uses the default-
filtered listing (4.03%) — confusingly different numbers for the same
concept. Move isEligibleVaultGroupMiniChartVault to the client-safe
top-vaults helpers (re-exported from the server module for the
chart-data endpoints) and compute the description yield over that set,
which matches both the mini chart and the table's default filters.
